fix(lazygit): check if default config file exists before adding to LG_CONFIG_FILE (#2256)
## Description

I noticed that lazygit was erroring out on open when the default config
file was missing even though lazygit outside of neovim opened fine
without it. Tracked it to the `LG_CONFIG_FILE` var that tells lazygit to
error out when any of the files are missing - and snacks was setting
this to include the default config even when it didn't exist.

## Related Issue(s)

fixes issue: https://github.com/folke/snacks.nvim/issues/1548