docs(rename): netrw example (#2153)
## Description

<!-- Describe the big picture of your changes to communicate to the
maintainers
  why we should accept this pull request. -->

I recently got really into improving my experience with netrw, in one of
those explorations I managed to integrate with the `Snacks.rename`
module by remapping the default rename mapping. So I thought it would be
cool to include a more native approach to the documentation of this
module.

> PS: In the example I implemented with the `vim.ui.input` interface
since it should be the new standard, but if you prefer with the original
`vim.fn.input` I can change it.

## Related Issue(s)

<!--
  If this PR fixes any issues, please link to the issue here.
  - Fixes #<issue_number>
-->

N/a

## Screenshots

<!-- Add screenshots of the changes if applicable. -->


[![asciicast](https://asciinema.org/a/735692.svg)](https://asciinema.org/a/735692)